Get GitHub Gist Urls

Given a username, return a dataframe of all the gists by that user. Note that (similar to other github_* functions) if run repeatedly may error due to hitting github API rate limits.

Usage

github_gists(user)

Arguments

user

Character.

Value

Dataframe with contents and urls of file paths.

Details

Returned output does not filter to only R and Rmarkdown files by default, see example for typical use with funspotr:::str_detect_r_rmd() -- I use a lot of .md files for gists so should probably write something to parse out the R code chunks from .md files.

Examples

library(dplyr)
library(funspotr)

# pulling and analyzing my R file github gists
gists_urls <- github_gists("brshallo") %>%
  filter(funspotr:::str_detect_r_rmd(contents))

gists_urls
#> # A tibble: 34 × 2
#>    contents                                  urls                               
#>    <chr>                                     <chr>                              
#>  1 custom-ggplot-and-labels.R                https://gist.githubusercontent.com…
#>  2 stratified-sampling-parameter-estimates.R https://gist.githubusercontent.com…
#>  3 cdf_density.R                             https://gist.githubusercontent.com…
#>  4 if_all-if_any-examples.R                  https://gist.githubusercontent.com…
#>  5 bootstrap-mean-rsample.R                  https://gist.githubusercontent.com…
#>  6 predict-interval-boot-only.R              https://gist.githubusercontent.com…
#>  7 source_rmd.R                              https://gist.githubusercontent.com…
#>  8 source-rmd-chunks.r                       https://gist.githubusercontent.com…
#>  9 predict-interval.R                        https://gist.githubusercontent.com…
#> 10 cut-floor-ceiling.R                       https://gist.githubusercontent.com…
#> # … with 24 more rows

# Will just parse the first 2 files/gists
contents <-
  github_spot_funs(custom_urls = slice(gists_urls, 1:2))
#> Warning: incomplete final line found on 'https://gist.githubusercontent.com/brshallo/78d0682d16b7eb09ee39caa4eba3fc06/raw/50856a8f1b3c8049b630b71e05b1ebf92a4d5267/custom-ggplot-and-labels.R'
#> Warning: incomplete final line found on 'https://gist.githubusercontent.com/brshallo/e031e73711afd2fff26636bd6091dd58/raw/e28d8a0d4880b7ac7c3a726bb272571c3a623eb5/stratified-sampling-parameter-estimates.R'

contents %>%
  unnest_github_results()
#> # A tibble: 26 × 5
#>    funs          pkgs      in_multiple_pkgs contents        urls                
#>    <chr>         <chr>     <lgl>            <chr>           <chr>               
#>  1 library       base      FALSE            custom-ggplot-… https://gist.github…
#>  2 ggplot        (unknown) FALSE            custom-ggplot-… https://gist.github…
#>  3 aes           (unknown) FALSE            custom-ggplot-… https://gist.github…
#>  4 geom_point    (unknown) FALSE            custom-ggplot-… https://gist.github…
#>  5 custom_ggplot (unknown) FALSE            custom-ggplot-… https://gist.github…
#>  6 ggplotly      (unknown) FALSE            custom-ggplot-… https://gist.github…
#>  7 library       base      FALSE            stratified-sam… https://gist.github…
#>  8 tibble        (unknown) FALSE            stratified-sam… https://gist.github…
#>  9 c             base      FALSE            stratified-sam… https://gist.github…
#> 10 rep           base      FALSE            stratified-sam… https://gist.github…
#> # … with 16 more rows