Spot Packages or Functions from Github Repository

github_spot_pkgs() : Spot all packages that show-up in R or Rmarkdown documents in the github repository. Essentially a wrapper for mapping spot_pkgs() across multiple files.

github_spot_pkgs(
  repo,
  branch = "main",
  ...,
  preview = FALSE,
  rmv_index = TRUE,
  custom_urls = NULL
)

github_spot_funs(
  repo,
  branch = "main",
  ...,
  preview = FALSE,
  rmv_index = TRUE,
  custom_urls = NULL
)

Arguments

repo

Github repository, e.g. "brshallo/feat-eng-lags-presentation"

branch

Branch of github repository, default is "main".

...

Arguments passed onto spot_*().

preview

Logical, if set to TRUE will return dataframe of urls that will be passed through spot_funs() or spot_pkgs() without parsing files. See README for example of how this can be combined with custom_urls to only parse a subset of files identified.

rmv_index

Logical, most repos containing blogdown sites will have an index.R file at the root. Change to TRUE if you don't want this file removed.

custom_urls

Option to pass in a dataframe with columns contents and urls to overrride the default urls in the repo to parse. Default is NULL. See README for example.

Value

Dataframe with contents and urls of file paths along with a list-column spotted containing purrr::safely() lists of "result" and "error" for each file parsed. See unnest_github_results() for helper to put into an easier to read format.

Details

github_spot_funs() : Spot all functions and their corresponding packages that show-up in R or Rmarkdown documents in the github repository. Essentially a wrapper for mapping spot_pkgs() across multiple files.

Meant for cases where packages and scripts are in the same file. In cases where this is not the case will need to build an alternative workflow. See unexported functions in R/github-spot.R for some potentially helpful building blocks.

See also

Examples

library(funspotr)

github_spot_funs("brshallo/feat-eng-lags-presentation", branch = "main")
#> # A tibble: 4 × 3
#>   contents                      urls                                  spotted   
#>   <chr>                         <chr>                                 <list>    
#> 1 R/Rmd-to-R.R                  https://raw.githubusercontent.com/br… <named li…
#> 2 R/feat-engineering-lags.R     https://raw.githubusercontent.com/br… <named li…
#> 3 R/load-inspections-save-csv.R https://raw.githubusercontent.com/br… <named li…
#> 4 R/types-of-splits.R           https://raw.githubusercontent.com/br… <named li…

# Say you only want to parse a subset of the R file discovered, you could do
# something like:
library(dplyr)

github_spot_funs("brshallo/feat-eng-lags-presentation", branch = "main", preview = TRUE) %>%
  slice(1:2) %>%
  github_spot_funs(custom_urls = .)
#> # A tibble: 2 × 3
#>   contents                  urls                                     spotted    
#>   <chr>                     <chr>                                    <list>     
#> 1 R/Rmd-to-R.R              https://raw.githubusercontent.com/brsha… <named lis…
#> 2 R/feat-engineering-lags.R https://raw.githubusercontent.com/brsha… <named lis…