funspotr Network Plot

Output simple network plot using visNetwork connecting either funs or pkgs to relative_paths/absolute_paths.

Usage

network_plot(df, to = .data$pkgs, show_each_use = FALSE)

Arguments

df

Dataframe containing columns relative_paths, absolute_paths and either funs or pkgs. Generally the output from running: github_spot_*() %>% unnest_results()

to

funs or pkgs

show_each_use

Binary, default is FALSE. If TRUE edge thickness will be based on the number of times a package or function is used.

Value

visNetwork plot

Examples

# \donttest{
library(dplyr)
library(funspotr)

gh_ex_pkgs <- list_files_github_repo(
  repo = "brshallo/feat-eng-lags-presentation",
  branch = "main") %>%
  spot_funs_files()

gh_ex_pkgs %>%
  unnest_results() %>%
  network_plot(to = pkgs)

# }